requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
click==8.1.7
colorama==0.4.6
coverage==7.3.2
cssbeautifier==1.14.9
Django==4.2.6
django-appconf==1.0.5
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
django-imagekit==5.0.0
django-sendfile2==0.7.0
djangorestframework==3.14.0
djlint==1.34.0
drf-nested-routers==0.93.4
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.9
json5==0.9.14
pathspec==0.11.2
pilkit==3.0
Pillow==10.0.1
pkg_resources==0.0.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
